pyproject.toml
[tool.poetry]
name = "uca-backend"
version = "0.1.0"
description = ""
authors = ["Tom Smith <thomas-smith@sky.com>"]
readme = "README.md"
packages = [{include = "uca_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
azure-cosmos = "^4.3.0"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
autoflake = "^1.7.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
